+++ src/pysharpe/optimization/__init__.py
@@ -1,5 +1,6 @@
 """Optimisation helpers."""
 
+from .bayesian import BayesianOptimizer
 from .models import OptimisationPerformance, OptimisationResult, PortfolioWeights
 from .weights import normalize_weights
 
@@ -8,4 +9,5 @@ __all__ = [
     "OptimisationPerformance",
     "OptimisationResult",
     "normalize_weights",
+    "BayesianOptimizer",
 ]